backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# ML/Training
torch>=2.1.0
transformers>=4.36.0
datasets>=2.15.0
peft>=0.7.0
trl>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
axolotl>=0.4.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
jsonschema>=4.20.0
pyyaml>=6.0.1
openapi-spec-validator>=0.7.1

# Utilities
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
pytest>=7.4.0
pytest-django>=4.7.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0